Add an option to start miners with a Normal ProcessWindowStyle. Miners are now started without stealing focus from the user.
A optional parameter, "-minervisibility", is added, which defaults to "minimized". The two options are "normal" and "minimized". A "hidden" option was tested, but it caused miner processes to not stop properly for an unknown reason. "-minervisibility" is kept as an enum instead of a switch so that a "hidden" option can easily be added in the future.

Start-SubProcess was altered to use CreateProcess from the Windows API instead of Start-Process built into PowerShell. This was necessary to allow miner processes to be started without stealing focus.

Updated README.md and README.txt to reflect the above changes.