Fix for log file getting corrupted
It seems to be a combination of two issues:

One is Out-File sometimes deciding the file is Unicode instead of ASCII,
and writing with a space (actually a NUL) between each character (see
https://stackoverflow.com/questions/11147179/nul-byte-between-every-other-character-in-output)

The other is a race condition that happens very rarely, corrupting the log file when multiple
threads/processes/jobs try to write to the file at exactly the same moment. This one is
fixed using mutexes, which are shared between processes, and cause the
program to wait up to 1 second for any other programs to release the
mutex before writing to the file. In reality, the wait will be much
shorter than that, usually 0.  The timeout is just to ensure the program
doesn't get stuck there if somehow the mutex didn't get released.