Use same verbosity for Write-Log as main script
Because it's in a separate module, Write-Log doesn't inherit the
verbose, debug and info preferences of the main script.  This sets them
to be the same as the main script.

Makes debugging easier when running MultiPoolMiner.ps1 with -debug
and/or -verbose.  No noticable change for anyone running the script
normally.