Add Enemy 1.08 Miner by Zealot
This miner does not have a GitHub page and has a Mega link instead. Dont know how you guys feel about these...
Please take a look at the algos structure. I started with the Tpruvot structure and tested all algos. Ive commented out the ones that are not supported. There are a few I couldn't test and noted that in the comment, but didn't want to prevent testing in case a pool supports them at some point. 
If you guys like it I can copy that algo structure to all ccminer forks and comment out the unsupported algos for each miner. I also think that renaming the miners to ccminerDevName would help with updates and adding future better miners which could outperform the ones that are algo specific for example. Let me know what you think and I can do that while going through the miners. 
This miner has a 1% fee and I know MPM takes that into account. I remember seeing it into the code, but couldn't find it when i needed it... If someone can help with that I'd appreciate it.