Set current path used by .NET methods
Sometimes [IO.Path] calls will see the current directory as the user's
home directory (eg. C:\user\username) instead of the current powershell
directory.  .NET has a separate variable for the current directory.

That was causing it to extract files to the wrong location.

This makes sure it matches the script's current directory.