Fix parameter '-PoolName'
Parameter '-PoolName' was not working under the following scenario:

- Configfile contains hardcoded Pool settings ($Config.Pools.[PoolName])

In this configuration the 'PoolName' setting (either passed as command line option or hard coded in the config file) is without effect. 
MPM would use ALL pools for which a Config.Pools.[PoolName] exists regardless of what the value of 'Poolname / $Config.PoolName' was.